Move all the caches into their own package, synapse.util.caches
