+++ synapse/storage/stream.py
@@ -35,7 +35,8 @@ what sort order was used:
 
 from twisted.internet import defer
 
-from ._base import SQLBaseStore, cachedInlineCallbacks
+from ._base import SQLBaseStore
+from synapse.util.caches.descriptors import cachedInlineCallbacks
 from synapse.api.constants import EventTypes
 from synapse.types import RoomStreamToken
 from synapse.util.logutils import log_function
